Add a helper for formatted printing (#18)
Used for error messages now, and will help with serial output for
tracing, later.

+++ src/chip8/state.hpp
@@ -30,7 +30,7 @@ struct EmuState {
     uint16_t Stack[16] = {0};
     
     // Stack Pointer
-    uint16_t StackPointer = 0;
+    uint8_t StackPointer = 0;
 
     // ArduboyChip8 data
     // True if the emualtor is in the running state
